Marker position not updated if metadata already loaded.
If the markers were created after the video metadata was loaded the
onMetadaLoaded event was never fired and the markers where not
positioned.

I added a check for readyState greater than 0 (metadata loaded). In that
case the update is performed.

The function is also called the next time the metadata is loaded. This
will happen if meta data was not loaded (as with the previous code),
but also in case that there has been a change of video source and the
markers were positioned according to the old source duration.

Changes to be committed:
	modified:   src/markerBar.js